rename: RecruitmentParticipation 도메인 패키지명 변경
