+++ src/main/java/project/volunteer/DummyConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import project.volunteer.domain.image.dao.ImageRepository;
-import project.volunteer.domain.participation.dao.ParticipantRepository;
+import project.volunteer.domain.recruitmentParticipation.repository.ParticipantRepository;
 import project.volunteer.domain.recruitment.repository.RecruitmentRepository;
 import project.volunteer.domain.recruitment.repository.RepeatPeriodRepository;
 import project.volunteer.domain.scheduleParticipation.repository.ScheduleParticipationRepository;
