+++ src/main/java/project/volunteer/domain/recruitmentParticipation/application/ParticipationService.java
@@ -1,7 +1,7 @@
-package project.volunteer.domain.participation.application;
+package project.volunteer.domain.recruitmentParticipation.application;
 
-import project.volunteer.domain.participation.application.dto.AllParticipantDetails;
-import project.volunteer.domain.participation.domain.Participant;
+import project.volunteer.domain.recruitmentParticipation.application.dto.AllParticipantDetails;
+import project.volunteer.domain.recruitmentParticipation.domain.Participant;
 import project.volunteer.domain.recruitment.domain.Recruitment;
 import project.volunteer.domain.user.domain.User;
