+++ src/main/java/project/volunteer/domain/scheduleParticipation/domain/ScheduleParticipation.java
@@ -3,8 +3,8 @@ package project.volunteer.domain.scheduleParticipation.domain;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import project.volunteer.domain.participation.converter.StateConverter;
-import project.volunteer.domain.participation.domain.Participant;
+import project.volunteer.domain.recruitmentParticipation.converter.StateConverter;
+import project.volunteer.domain.recruitmentParticipation.domain.Participant;
 import project.volunteer.domain.sehedule.domain.Schedule;
 import project.volunteer.global.common.auditing.BaseEntity;
 import project.volunteer.global.common.component.ParticipantState;
