+++ src/main/java/project/volunteer/domain/scheduleParticipation/mapper/ScheduleParticipantFacade.java
@@ -3,8 +3,8 @@ package project.volunteer.domain.scheduleParticipation.mapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import project.volunteer.domain.participation.application.ParticipationService;
-import project.volunteer.domain.participation.domain.Participant;
+import project.volunteer.domain.recruitmentParticipation.application.ParticipationService;
+import project.volunteer.domain.recruitmentParticipation.domain.Participant;
 import project.volunteer.domain.scheduleParticipation.service.ScheduleParticipationDtoService;
 import project.volunteer.domain.scheduleParticipation.service.ScheduleParticipationService;
 import project.volunteer.domain.scheduleParticipation.service.dto.CancelledParticipantList;
