+++ src/main/java/project/volunteer/domain/scheduleParticipation/service/ScheduleParticipationService.java
@@ -1,6 +1,6 @@
 package project.volunteer.domain.scheduleParticipation.service;
 
-import project.volunteer.domain.participation.domain.Participant;
+import project.volunteer.domain.recruitmentParticipation.domain.Participant;
 import project.volunteer.domain.sehedule.domain.Schedule;
 
 import java.util.List;
