+++ src/main/java/project/volunteer/global/error/exhadler/ParticipationControllerAdvice.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import project.volunteer.domain.participation.api.ParticipationController;
+import project.volunteer.domain.recruitmentParticipation.api.ParticipationController;
 import project.volunteer.global.error.exception.BusinessException;
 import project.volunteer.global.error.exception.ErrorCode;
 import project.volunteer.global.error.response.BaseErrorResponse;
