+++ src/main/java/project/volunteer/global/error/exhadler/UserControllerAdvice.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import project.volunteer.domain.participation.api.ParticipationController;
 import project.volunteer.domain.user.api.UserController;
 import project.volunteer.global.error.exception.BusinessException;
 import project.volunteer.global.error.exception.ErrorCode;
