+++ src/test/java/project/volunteer/domain/logboard/api/LogboardWriteControllerTest.java
@@ -41,8 +41,8 @@ import project.volunteer.domain.image.dao.ImageRepository;
 import project.volunteer.domain.image.dao.StorageRepository;
 import project.volunteer.domain.image.domain.Storage;
 import project.volunteer.domain.logboard.domain.Logboard;
-import project.volunteer.domain.participation.dao.ParticipantRepository;
-import project.volunteer.domain.participation.domain.Participant;
+import project.volunteer.domain.recruitmentParticipation.repository.ParticipantRepository;
+import project.volunteer.domain.recruitmentParticipation.domain.Participant;
 import project.volunteer.domain.recruitment.application.RecruitmentCommandUseCase;
 import project.volunteer.domain.recruitment.repository.RecruitmentRepository;
 import project.volunteer.domain.recruitment.domain.Recruitment;
